Workaround for Xft crash when certain glyphs are rendered wtf..
For example rendering a Thin Space (U+2009) in terminus-font causes a
crash on NixOS. The glyph is present and set to all blank pixels. Other
glyphs have an identical representation (such as ascii space) and they
don't cause the crash. No idea what the problem is but for now we'll
just hack the programs that render arbitrary text with Xft (just dwm and
st - dmenu shouldn't be affected since it only shows program names) so
that they never draw a Thin Space in terminus.