Fix emacs fonts in Xdefaults
Ideally the font setting for emacs in Xdefaults would be moved out of
this repo since it's not portable across systems. Currently I only
have a single machine running X  where I use emacs but if this was to
change then the Xdefaults file checked in here would probably not
work (unless it ran the same version of NixOS).